refactor: change folders names
